Rename and move GUI packages around.
