+++ src/main/java/fiji/plugin/trackmate/LoadTrackMatePlugIn.java
@@ -10,7 +10,7 @@ import javax.swing.JFrame;
 import org.scijava.util.VersionUtils;
 
 import fiji.plugin.trackmate.gui.GuiUtils;
-import fiji.plugin.trackmate.gui.LogPanel;
+import fiji.plugin.trackmate.gui.components.LogPanel;
 import fiji.plugin.trackmate.gui.displaysettings.DisplaySettings;
 import fiji.plugin.trackmate.gui.wizard.TrackMateWizardSequence;
 import fiji.plugin.trackmate.gui.wizard.descriptors.ConfigureViewsDescriptor;
