+++ src/main/java/fiji/plugin/trackmate/detection/SpotDetectorFactoryBase.java
@@ -8,7 +8,7 @@ import fiji.plugin.trackmate.Model;
 import fiji.plugin.trackmate.Settings;
 import fiji.plugin.trackmate.SpotRoi;
 import fiji.plugin.trackmate.TrackMateModule;
-import fiji.plugin.trackmate.gui.ConfigurationPanel;
+import fiji.plugin.trackmate.gui.components.ConfigurationPanel;
 import net.imagej.ImgPlus;
 import net.imglib2.type.NativeType;
 import net.imglib2.type.numeric.RealType;
