+++ src/main/java/fiji/plugin/trackmate/gui/displaysettings/StyleElements.java
@@ -39,10 +39,10 @@ import javax.swing.border.EmptyBorder;
 import org.drjekyll.fontchooser.FontDialog;
 
 import fiji.plugin.trackmate.Settings;
-import fiji.plugin.trackmate.gui.FeatureDisplaySelector;
 import fiji.plugin.trackmate.gui.GuiUtils;
+import fiji.plugin.trackmate.gui.components.CategoryJComboBox;
+import fiji.plugin.trackmate.gui.components.FeatureDisplaySelector;
 import fiji.plugin.trackmate.gui.displaysettings.DisplaySettings.TrackMateObject;
-import fiji.plugin.trackmate.gui.panels.components.CategoryJComboBox;
 
 public class StyleElements
 {
