+++ src/main/java/fiji/plugin/trackmate/gui/wizard/descriptors/ActionChooserDescriptor.java
@@ -2,9 +2,9 @@ package fiji.plugin.trackmate.gui.wizard.descriptors;
 
 import fiji.plugin.trackmate.SelectionModel;
 import fiji.plugin.trackmate.TrackMate;
+import fiji.plugin.trackmate.gui.components.ActionChooserPanel;
 import fiji.plugin.trackmate.gui.displaysettings.DisplaySettings;
 import fiji.plugin.trackmate.gui.wizard.WizardPanelDescriptor2;
-import fiji.plugin.trackmate.gui.wizard.descriptors.panels.ActionChooserPanel;
 import fiji.plugin.trackmate.providers.ActionProvider;
 
 public class ActionChooserDescriptor extends WizardPanelDescriptor2
