+++ src/main/java/fiji/plugin/trackmate/gui/wizard/descriptors/ExecuteDetectionDescriptor.java
@@ -3,7 +3,7 @@ package fiji.plugin.trackmate.gui.wizard.descriptors;
 import org.scijava.Cancelable;
 
 import fiji.plugin.trackmate.TrackMate;
-import fiji.plugin.trackmate.gui.LogPanel;
+import fiji.plugin.trackmate.gui.components.LogPanel;
 import fiji.plugin.trackmate.gui.wizard.WizardPanelDescriptor2;
 
 public class ExecuteDetectionDescriptor extends WizardPanelDescriptor2
