+++ src/main/java/fiji/plugin/trackmate/gui/wizard/descriptors/LogPanelDescriptor2.java
@@ -1,6 +1,6 @@
 package fiji.plugin.trackmate.gui.wizard.descriptors;
 
-import fiji.plugin.trackmate.gui.LogPanel;
+import fiji.plugin.trackmate.gui.components.LogPanel;
 import fiji.plugin.trackmate.gui.wizard.WizardPanelDescriptor2;
 
 public class LogPanelDescriptor2 extends WizardPanelDescriptor2
