+++ src/main/java/fiji/plugin/trackmate/gui/wizard/descriptors/SaveDescriptor.java
@@ -9,7 +9,7 @@ import javax.swing.SwingUtilities;
 
 import fiji.plugin.trackmate.Logger;
 import fiji.plugin.trackmate.TrackMate;
-import fiji.plugin.trackmate.gui.LogPanel;
+import fiji.plugin.trackmate.gui.components.LogPanel;
 import fiji.plugin.trackmate.gui.displaysettings.DisplaySettings;
 import fiji.plugin.trackmate.gui.wizard.WizardSequence;
 import fiji.plugin.trackmate.io.IOUtils;
