+++ src/main/java/fiji/plugin/trackmate/gui/wizard/descriptors/SomeDialogDescriptor.java
@@ -2,7 +2,7 @@ package fiji.plugin.trackmate.gui.wizard.descriptors;
 
 import java.io.File;
 
-import fiji.plugin.trackmate.gui.LogPanel;
+import fiji.plugin.trackmate.gui.components.LogPanel;
 import fiji.plugin.trackmate.gui.wizard.WizardPanelDescriptor2;
 
 /**
