+++ src/main/java/fiji/plugin/trackmate/tracking/SpotTrackerFactory.java
@@ -7,7 +7,7 @@ import org.jdom2.Element;
 import fiji.plugin.trackmate.Model;
 import fiji.plugin.trackmate.SpotCollection;
 import fiji.plugin.trackmate.TrackMateModule;
-import fiji.plugin.trackmate.gui.ConfigurationPanel;
+import fiji.plugin.trackmate.gui.components.ConfigurationPanel;
 
 public interface SpotTrackerFactory extends TrackMateModule
 {
