+++ src/main/java/fiji/plugin/trackmate/tracking/overlap/OverlapTrackerSettingsPanel.java
@@ -22,7 +22,7 @@ import javax.swing.JPanel;
 import javax.swing.JRadioButton;
 import javax.swing.SwingConstants;
 
-import fiji.plugin.trackmate.gui.ConfigurationPanel;
+import fiji.plugin.trackmate.gui.components.ConfigurationPanel;
 import fiji.plugin.trackmate.tracking.overlap.OverlapTracker.IoUCalculation;
 
 public class OverlapTrackerSettingsPanel extends ConfigurationPanel
