+++ src/main/java/fiji/plugin/trackmate/tracking/sparselap/SimpleSparseLAPTrackerFactory.java
@@ -3,8 +3,8 @@ package fiji.plugin.trackmate.tracking.sparselap;
 import org.scijava.plugin.Plugin;
 
 import fiji.plugin.trackmate.Model;
-import fiji.plugin.trackmate.gui.ConfigurationPanel;
-import fiji.plugin.trackmate.gui.panels.tracker.SimpleLAPTrackerSettingsPanel;
+import fiji.plugin.trackmate.gui.components.ConfigurationPanel;
+import fiji.plugin.trackmate.gui.components.tracker.SimpleLAPTrackerSettingsPanel;
 import fiji.plugin.trackmate.tracking.SpotTrackerFactory;
 
 @Plugin( type = SpotTrackerFactory.class )
